修复重大bug 一定一注意 是new char[], 不要错误的写成new char()

+++ Chat/Chat/NetWork/msgstructure.h
@@ -59,7 +59,7 @@ typedef struct _ResponseLoginMsg {
 
 //friendlist response
 typedef struct _ResponseFriendList {
-    long len;
+    int len;
     char friendlist[0];
 }ResponseFriendList;
